feat: remove local deployment and updates
