--- packages/votingV2/docker-compose.yml
@@ -1,45 +1,0 @@
-version: '3'
-services:
-  graph-node:
-    image: graphprotocol/graph-node
-    ports:
-      - '8000:8000'
-      - '8001:8001'
-      - '8020:8020'
-      - '8030:8030'
-      - '8040:8040'
-    depends_on:
-      - ipfs
-      - postgres
-    extra_hosts:
-      - 172.18.0.1:host-gateway
-    environment:
-      postgres_host: postgres
-      postgres_user: graph-node
-      postgres_pass: let-me-in
-      postgres_db: graph-node
-      ipfs: 'ipfs:5001'
-      ethereum: 'goerli:https://goerli.infura.io/v3/${INFURA_API_KEY}'
-      GRAPH_LOG: info
-  ipfs:
-    image: ipfs/go-ipfs:v0.10.0
-    ports:
-      - '5001:5001'
-    volumes:
-      - ./data/ipfs:/data/ipfs
-  postgres:
-    image: postgres
-    ports:
-      - '5432:5432'
-    command:
-      [
-        "postgres",
-        "-cshared_preload_libraries=pg_stat_statements"
-      ]
-    environment:
-      POSTGRES_USER: graph-node
-      POSTGRES_PASSWORD: let-me-in
-      POSTGRES_DB: graph-node
-      PGDATA: "/data/postgres"
-    volumes:
-      - ./data/postgres:/var/lib/postgresql/data
